requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asteroid-filterbanks==0.4.0
attrs==26.1.0
audioop-lts==0.2.2
audioread==3.1.0
brotli==1.2.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorlog==6.10.1
contourpy==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.4
cuda-toolkit==13.0.2
cycler==0.12.1
decorator==5.2.1
einops==0.8.2
fastapi==0.136.1
filelock==3.29.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.4.0
googleapis-common-protos==1.74.0
gradio==6.14.0
gradio-client==2.5.0
greenlet==3.5.0
groovy==0.1.2
grpcio==1.80.0
h11==0.16.0
hf-gradio==0.4.1
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.2
idna==3.13
importlib-metadata==8.7.1
jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.47.0
mako==1.3.12
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
more-itertools==11.0.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
onnxruntime==1.25.1
openai-whisper==20250625
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-exporter-otlp-proto-http==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
optuna==4.8.0
orjson==3.11.8
packaging==26.2
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
primepy==1.3
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyannote-audio==4.0.4
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pycparser==3.0
pydantic==2.13.3
pydantic-core==2.46.3
pydub==0.25.1
pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.27
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
pytz==2026.2
pyyaml==6.0.3
qwen-tts==0.1.1
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.1.0
sqlalchemy==2.0.49
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.11.0
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.11.0
torchcodec==0.11.1
torchmetrics==1.9.0
tqdm==4.67.3
transformers==4.57.3
triton==3.6.0
typer==0.25.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.46.0
yarl==1.23.0
zipp==3.23.1
